Fix bug in test suite
TH declarations have a very weird property: the names in

```haskell
[d| data T = MkT { foo :: .. , .. } |]
```

have a `Uniq` associated with them (like `newName`), rather than being
dynamically bound (like `mkName`). This makes no sense but is what happens, and
I had adjusted the name treatment to allow for it. But we were using some TH
generation for large records in the test suite that weren't doing this, and
this caused trouble.