Make the operations on Rep less strict
The key piece of the puzzle here is what is now exported as the
"array-like" interface. In order to support, for example,

```haskell
Rep.zipWith .. undefined undefined
```

and have it result in something useable, we use `allIndices` to
statically compute the shape of the result, and then use `getByIndex` to
index the input vectors, but only if the specified function demands it.
Since `Rep` is internally represented as a vector with O(1) access, this
is not actually slower.