Refactor treatment of names
Although I tried to be careful in how we deal with names, it will had
some deficiencies. In particular, it was difficult to share naming
strategies between TH and QQ. This commit significantly revises the
naming infrastructure, improves sharing between TH and QQ, and undoes
some weirdness that was introduced by TH itself.