Start using vector as internal representation
The original goal of this commit was simply to add the constrained
combinators. However, while doing so, I realised that this was not
possible with the existing internal record representation: we had no
information about the _order_ of the fields, and we really should
traverse records in row order for monadic combinators.

Thus started a pretty major refactoring. The internal representation is
now vector based, with the fields of the vector guaranteed to be in row
order. This means that `mapM` and friends can just map over the vector,
and will do the right thing. We address the complexity problem arising
from updates/inserts to vectors by representing a record as a vector and
a diff; see `Diff` for details.

I took the opportunity to cleanup the library, split things into smaller
modules, be careful about namespacing, and introduce the `.Advanced`
top-level module (which will get a `.Simple` cousin, which will be
`Record` without the functor argument).

Some notable changes that aren't just refactoring:

- The set of combinators is now complete, and have corresponding
  property tests.
- Added `Generic` sanity check for functor other than `I`.
- The `RecordMetadata` class has been replaced by a more primitive
  `KnownFields` class. Importantly, the new class does not have a
  functor argument.
- The property tests now specifically test for records with fields in
  different order, to test that the monadic combinators respect this
  order.

This commit also adds some TODOs:

- We should be able to prove `AllFields` from a vector of dictionaries
  (i.e., in the "reverse direction").
- Define `.Simple` module.
- `apply :: Diff f -> ..` should get some special cases.
- We might be able to give `evidenceHasField` an actual index (as well
  as field name). After all, we only resolve this when we know the
  concrete (prefix of) the record.
- We should restore the doctests; currently disabled because they don't
  work so well with the new module structure.