Introduce GHC.TemplateHaskellStyle
The original PR introduced many functions that mirrored the TH interface
for the GHC API, but it wasn't entirely consistent about it. I really
like this idea, so I've pushed it quite a bit further, and now the
entire codebase is defined in terms of this new modulo.

The most important thing still not quite right is that we are losing
location information.