Delete `canonIndices` from `Canonical`
Instead we compute all indices at runtime. We only _use_ them if we have
no `Diff`, but there is no need to do a `HashMap` lookup anymore on the
field names. This means that field lookup will be `O(1)` in the typical
case, and `O(log n)` in the worst case (if we have a `Diff`).