Extend SR benchmarks to 80 fields
Using the unsafe API allows us to do so.

This also fixes the measurement tools to close each file after we
process it, to avoid running out of file handles.

We only show -ddump-simpl for superrecord to avoid the core size
measurements taking to look. The full benchmark suite now takes roughly
25 minutes on my machine.