Remove annotation after processing large record
It's possible for the compiler plugin to fire multiple times, with
each successive run using the AST from the previous run.

When this happens, the first run succeeds but later runs fail. The
datatype definition generated by `transformDecl` still has the
`largeRecord` annotation, but is no longer in a form that's accepted
by the compiler plugin.

Removing the annotations for which `transformDecl` succeeds makes the
compiler plugin idempotent.