Separate HsSrcBang out of field type
The main motivation is that GHC-9.14 forces us to things this way.
And even if it didn't, this makes sense.
"Bangs" are not part of the type, they are part of the field
declaration, and once we already keep them separate (e.g. in Field),
it's good to simply never mix them.